transformer refactor/further debug
